+++ samples/boards/nrf52/mesh/onoff_level_lighting_vnd_app/src/app_gpio.h
@@ -5,9 +5,13 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#ifndef _COMMON_H
-#define _COMMON_H
+#ifndef _APP_GPIO_H
+#define _APP_GPIO_H
 
-void update_light_state(void);
+/* GPIO */
+extern struct device *button_device[4];
+extern struct device *led_device[4];
+
+void app_gpio_init(void);
 
 #endif
